fix #64: Исправлена ошибка обязательности HEADERS и баг с транзакциями mysql

+++ lib/transport/bitrix/connection.php
@@ -3,6 +3,8 @@
 namespace Bsi\Queue\Transport\Bitrix;
 
 use Bitrix\Main\Application;
+use Bitrix\Main\DB\MysqlCommonConnection;
+use Bitrix\Main\DB\SqlQueryException;
 use Bitrix\Main\ORM\Query\Query;
 use Bitrix\Main\Type\DateTime;
 use Symfony\Component\Messenger\Exception\InvalidArgumentException;
@@ -28,9 +30,12 @@ class Connection
      */
     protected $configuration = [];
 
+    private $doMysqlCleanup;
+
     public function __construct(array $configuration = [])
     {
         $this->configuration = array_replace_recursive(static::DEFAULT_OPTIONS, $configuration);
+        $this->doMysqlCleanup = false;
     }
 
     public static function buildConfiguration(string $dsn, array $options = []): array
@@ -85,6 +90,18 @@ class Connection
     {
         $driverConnection = Application::getConnection(MessageTable::getConnectionName());
 
+        if ($this->doMysqlCleanup && $this->isMysqlConnection()) {
+            try {
+                $driverConnection->query(sprintf(
+                    "DELETE FROM %s WHERE DELIVERED_AT = '9999-12-31 23:59:59'",
+                    MessageTable::getTableName(),
+                ));
+                $this->doMysqlCleanup = false;
+            } catch (SqlQueryException $e) {
+                // Ignore the exception
+            }
+        }
+
         $driverConnection->startTransaction();
         try {
             $query = $this->createAvailableMessagesQuery()
@@ -94,6 +111,7 @@ class Connection
             $bitrixEnvelope = $query->exec()->fetch();
 
             if ($bitrixEnvelope === false || $bitrixEnvelope === null) {
+                $driverConnection->commitTransaction();
                 return null;
             }
 
@@ -113,6 +131,17 @@ class Connection
 
     public function ack(int $id): bool
     {
+        if ($this->isMysqlConnection()) {
+            $result = MessageTable::update($id, ['DELIVERED_AT' => new DateTime('9999-12-31 23:59:59', 'Y-m-d H:i:s')]);
+            $updated = $result->isSuccess();
+
+            if ($updated) {
+                $this->doMysqlCleanup = true;
+            }
+
+            return $updated;
+        }
+
         $result = MessageTable::delete($id);
         if (!$result->isSuccess()) {
             throw new TransportException(implode("\n", $result->getErrorMessages()), 0);
@@ -166,5 +195,10 @@ class Connection
             )
             ->where('AVAILABLE_AT', '<=', $now)
             ->where('QUEUE_NAME', $this->configuration['queue_name']);
+    }
+
+    private function isMysqlConnection(): bool
+    {
+        return Application::getConnection(MessageTable::getConnectionName()) instanceof MysqlCommonConnection;
     }
 }
